Updating to use Makefile and updating path
When running on Fedora 21, the path to openshift is in the following location:

    _output/local/bin/linux/amd64/openshift

Not sure if this is across all builds though.  If not, I can update the documentation accordingly or add an alias.

Also, the Makefile usage for calling clean and then build can remove old artifacts from old builds that hack/build-go.sh wouldn't catch.