Properly version Docker image metadata sent to the API
Create a new internal object image/api.DockerImage which has two
versions, "1.0" and "pre012" to match the docker client and docker
registry API versioning.  Allow the version of that API to be
passed when submitting an image - if not passed, assume that the
object is "1.0" (which is our legacy behavior).  Clients should
set "apiVersion" inside the "DockerImageMetadata" field for maximum
forward compatibility to "1.0".

TODO: Upstream should allow the ability to decode and default a
kind and apiVersion for maximum flexibility.