Define "to" and "dockerImageReference" as new fields on BuildOutput
Instead of using ImageTag and Registry, which must be joined together,
allow clients to provide one of two options. "to", which may be a
reference to an ImageRepository, or "dockerImageReference", which is
a Docker push spec (the value you would tag an image before docker
pushing it to a registry).  Support backwards compatibility for ImageTag
and Registry.

Fix versioning of "BUILD" output - the field must be to a locked API
version so that an upgrade during a build doesn't break it.

Add proper validations on Build and BuildConfig, so that names,
namespaces, and labels validate according to Kubernetes rules, and
also increase the validation around the two new fields.