Vagrantfile: improve providers and usability/readability
OpenStack provider is now usable.

Refactored some things to hopefully improve the user's understanding of
what is going on with the various providers and options.

Choice of single/multiple hosts was separated from provider, because
even though multiple won't work with anything but VirtualBox, these
concepts are separate, and the error message was not so helpful if you
tried to request multiple with another provider.

Modified hack/vm-provision-full.sh such that it doesn't assume the
"vagrant" user so that images with other ssh-users can also get set
up properly.