Fix the Vagrant network setup
The Vagrantfile configures the machines with:

    master_ip = "10.245.1.2"
    minion_ip_base = "10.245.2."

The issue is that, by default, Vagrant uses a netmask of 255.255.255.0.
With that netmask, Vagrant considers that the master and the minions are on
two distinct subnets and so, creates two distinct private networks.
This prevents the master and the minions from communicating between each other.

By explicitly forcing the netmask to be 255.255.0.0, Vagrant will create a
single private network for the master and the minions.