Fix the Vagrant network setup
In order to have the minions able to communicate with the master, they must be
on the same private network. Vagrant considers by default that IPs are on
/24 networks. So, the IP of the master must be in the same /24 block than the
IP on the minions.

Vagrant complains when trying to assign an IP which ends with .1
So, 10.245.2.2 is the IP of the master,
10.245.2.3 and onwards are the IP of the minions.