Provide both a host and guest profile
openshift-node-host and openshift-node-guest profiles are to be used
for bare metal and virtual machine installations respectively. The rpm
installation scripts determine which profile is most appropriate based
on the output of the 'tuned-adm recommend' command so the user should
not need to make a selection on their own.

A manpage has been created as well.