[WIP] Use go instead of make
Stop using make instead of the go tool that is already used for
everything else. Things that need to be fixed before considering
merging this are noted as TODO comments inside the code.

Currently working scenarios include the simplest:
* Build code without using any flags: go run make.go
* Remove previous build: go run make.go --clean
* Build and run OpenShift all-in-one server: go run make.go --run

Note though that we currently don't get any output in stdout so
we won't get any of it in the last command as well.