Add service dependency triggering on enablement
This changes the point at which dependency triggers are created such
that when you enable a service it adds itself as a dependency rather
than when the package is installed. This allows for hosts with all
packages to be installed but for openshift-sdn-node to only be
triggered when you start openshift-node if openshift-sdn-node has been
enabled.