Add an edit command
`osc edit` allows any resource on the server to be edited directly from
the CLI, which greatly simplifies the process of making changes to objects.
The edited object will be the API object in YAML format by default.

Errors will be captured to the comment at the top of the file and then
output to the user. Validation errors will allow the user to attempt
to edit the file a second time.