Add osc stop and osc label
This commit adds two commands from Kubernetes: stop and label. While label was rather trivial to import in OpenShift, for stop to be fully functional deploymentConfigs had first to implement the kubectl.Resizer interface (work on the Resizer can be found in #2158) which is a prerequisite for another needed interface, kubectl.Reaper (work on the Reaper can be found here).