Expose routes from services
This commit adds the expose command from Kubernetes with a couple of additions. `osc expose` like `kubectl expose` exposes services, pods, or replication controllers as services and then it goes on exposing deployment configs as services and services as routes.

In order for `osc expose` to generate routes from services, routes had to implement the kubectl.Generator interface.