Add `osc env` command for setting and reading env
Allows osc users to easily add and remove environment variables
from pods, replication controllers, and deployment controllers.

Environment can be read from args, params, or STDIN. Environment
can be listed, modified in place and output, or modified in place
and sent to the server for update.

Adopts the initial "generation" pattern, which allows the input
to be decorated and then passed to a subsequent program for
further decoration.