Persist a master count config variable to enable multi-master
MasterCount is used by the master to determine the appropriate number
of masters in use so that the endpoints list can converge. Eventually,
we will have a better algorithm for master membership convergence.

The value will default to 1 and fail validation if < 1.  Admins must
set this to the same value on all masters.