Deployer pod failure detection improvements
When a pod is terminated by the kubelet in response to the deadline
being exceeded, there is no container status present to inspect on
the pod. The pod status will still be PodFailed.

Adjust the deployer pod failure detection to account for this. When
the pod status is PodSucceeded, use container status to decide if the
deployment failed. When the pod status is PodFailed, consider the
deployment a failure regardless of whether container status is available.