Controllers should be able to be lazily started
Allows HA master enablement after start. New endpoint
/controllers which supports GET (200 if controllers are
running, 202 if they are not, 405 if they can't be started),
PUT (200 if they are started, 405 if they can't be started),
and DELETE (200 causes server to shut down).

Add two new config values - Controllers (string with 'none'
and '*') which is the list of controllers to start, and
PauseControllers which if true waits for PUT /controllers
to start the second controller.

Allows external control of HA state