cli: Use ~/.kube/config instead of ~/.config/openshift/config
Kubernetes authentication is tricky enough as it is. We need kubectl,
oc, and oadmn to share authentication data, cluster context, namespace
configuration and so on.

People who build scripts around kubectl should have a reasonable expectation
of them working when they have logged into openshift via 'oc login'.
To make this happen, as well as other cases where config data is read
(such as Cockpit) the configuration (and especially auth/user/context) config
should live in one place, not two.

While ~/.config is nice and clean, the upstream path of ~/.kube carries
a lot of weight and standardizing on that seems most consistent and painless.