Show "no deployments" message on browse page consistently
The message was hidden, but nothing displayed when only replication
controllers with no `deploymentConfig` annotations existed. These can
be seen on the overview page, but are not displayed on the browse
deployments page.

Show the message when the page has no content and remove the message
when there is anything to show (including deployment configurations with
no deployments).

Fixes #3120
Fixes #2862