Fix volume dir label for e2e-docker
For test-end-to-end-docker.sh, use a volume directory under BASETMPDIR
and label it correctly.

The intermediate labeling solution is to chcon the volume dir on the host to
svirt_sandbox_file_t.

Longer term, we could

1. Update the selinux policy for volume dirs under /var/lib/openshift,
assuming that's the default location for volumes.
2. Modify
https://github.com/openshift/origin/blob/master/pkg/cmd/server/kubernetes/node.go#L102-L121
to always call chcon when the node starts, instead of only if the volume
dir doesn't exist and is therefore newly created.

This commit also changes the file name format of container logs captured
by the e2e tests. Instead of being container-$containerID.log, they are
now container-$podName-$containerName.log, such as
container-docker-registry-1-7sm9l-registry.log.