More e2e-docker fixes
Change BASETMPDIR to /tmp/openshift-e2e-containerized so it does not
share a common directory with test-end-to-end.sh. When they share a
common directory, if the volume mounts don't get cleaned up for whatever
reason, running the tests in succession can fail when they try to rm -rf
BASETMPDIR.

During test cleanup, attempt to delete all resources from the test and
default projects, and then wait up to 2 minutes for the volume mounts to
clear (it should only take approximately 10 seconds).