Prune blobs from registry even w/o stream refs
Update layer pruning to always delete blobs from the registry, even if
the image stream(s) that referenced the layer no longer exists. In the
event that there aren't any image streams referencing the layer any
more, we'll be able to delete the blob, but not the registry image
repository layer link files.

Fixes bug 1235148