Image pruning improvements
Update layer pruning to always delete blobs from the registry, even if
the image stream(s) that referenced the layer no longer exists. In the
event that there aren't any image streams referencing the layer any
more, we'll be able to delete the blob, but not the registry image
repository layer link files.

Fixes bug 1237271

Extract NewImagePruner arguments to ImagePrunerOptions struct.

Add attempt to check for connectivity to the registry before performing
actual pruning. If the connection attempt fails, report the error and
stop.

Reorganize image pruning operation order:

1. Remove image references from image streams
2. Remove registry data (repo layer links, blobs, repo manifest data)

If any errors occurred in 1 or 2, stop. Otherwise:

3. Remove images from OpenShift

This allows us to preserve images in the event there are issues deleting
from the registry. This way, additional calls to 'oadm prune' can be
attempted and we'll still have the image data to build the graph.