fedora 21 Vagrant provisioning fixes
1) Because udev renames network devices from their original enpXsY to
eth0 (since they are virtual and not hardware), the dracut-created
ifcfg-enp0s3 ifcfg file does not have a corresponding interface.  This
makes 'service network restart' or 'systemctl restart network' angry
and those commands terminate with an error.  Since the interface will
be automatically handled by NetworkManager anyway, we don't need the
ifcfg file so delete it.  Also, use nmcli to selectively start/stop
the 'eth1' interface since NetworkManager is handling it anyway.

2) Forces NFS shared directories when using libvirt, because it seems that
vagrant or vagrant-libvirt default to rsync for some reason, which of course
means that the OpenShift master generated certificates are not available for
the minions to use (since rsync is one-way).  This does mean that you need
to ensure that firewalld on your host lets NFS through:
    sudo firewall-cmd --permanent --add-service=nfs
    sudo firewall-cmd --permanent --add-service=rpc-bind
    sudo firewall-cmd --permanent --add-service=mountd
    sudo firewall-cmd --reload
See: http://nts.strzibny.name/vagrant-nfs-exports-on-fedora-21/

Other releases (like Centos7) may require NFS to use TCP since in those
releases firewalld blocks UDP packets by default.

Also make sure you have this patch to vagrant-libvirt:

https://github.com/pradels/vagrant-libvirt/commit/d107b899c8a3d0d22a9f8f7803f1e5753e4872e4

3) Finally, due to use of NFS, we need to ensure that hack/install-etcd.sh
doesn't try to use the owner from the original tarball (which is who-knows-what)
but instead uses the owner of the mounted NFS share.  This should be safe
for non-NFS and other Vagrant providers too.