Don't produce events when initial deployment fails
Don't produce events when initial deployment generation fails. The
most common case is when ImageChange triggers are defined for tags
which don't yet exist. Producing events in practice results in
thousands of duplicate events per deployment in some cases, and the
information is better relayed to users via new API state associated
with the trigger itself or the deployment config.