Vagrant: Allow override of IP addresses
master_ip and minion_ip_base were previously hard-coded in the
Vagrantfile, and the values corresponded with the subnet used by
virtualbox. On hosts that have both virtualbox and another provider
installed (e.g. vmware), attempting to deploy openshift with the
non-virtualbox provider is likely to result in VM's on the wrong
network. This change allows the addresses to be overridden via the
environment or configuration file.