Vagrant: Config synced folder type for devcluster
Allow the 'sync_folders_type' configuration value to set the type of the
shared folder (/vagrant) when deploying a devcluster.  This enables the
devcluster to be configured to share folders via nfs, which is faster
than the default shared folders used by virtualbox and vmware (libvirt
uses nfs by default).