Support election of controllers
A master can be configured to use a leader lease to determine
whether controllers can be started up. An admin would set
controllerLeaseTTL in master-config, and then start multiple
masters (or multiple controller processes). Each process would
try to acquire a leader lease on the key
<storageprefix>/leases/controllers, setting a TTL equal to the
configured value in seconds. Whichever controller acquires the
lock will be able to begin running. When a controller cannot
renew a lease due to errors talking to etcd, the key expires,
the key is deleted, or the value of the key is changed, the controller
will exit and terminate the process (either a full master, or just the
controller process).

Because terminating a master results in API connections being terminated
(including log sessions or exec proxy sessions) it is better to run the
masters as distinct API and controller sub processes.