Make output directory symlinks relative links
When these links contain the full paths of the source directory
then they are not relocatable. They should just refer to
the openshift binary in the same directory.

The most obvious case of this is when the directory that
openshift is built doesn't match the directory inside the
openshift-dev vagrant vm.