Enable connectivity between dind master and pods
The kube API supports querying a service and for this to work the master
needs to be able to reach the pods associated with a given service.
This change enables master to pod connectivity by running a node daemon
on the master (with schedulable=false) to ensure that the correct
routes are in place.

Since there are no longer different package requirements between a
master and node instance, the base, node and master docker images have
been collapsed into a single dind image.