Have routers take ownership of routes
When a router sees a route, it should make a decision about the name it
wishes to assign to that route (based on its override-hostname or own
settings) and then write the decision back to the route status with an
effective host value (so clients can see what the actual route value
is). If multiple routers attempt to make conflicting writes, have them
remember the conflict so as to avoid battles as they race to the new
value. Errors due to route uniqueness are also written back to the
status.

Each route has an array of RouteIngress structs, which contain an array
of conditions, the routerName (passed via --name to the router), and the
effective host.

Add printers and describers, make ingress empty be a special case.