Add a cluster diagnostic to check if master is also running as a node.
This is currently required a soft requirement for Open vSwitch to ensure the
master is on the SDN and can proxy to nodes.

Diagnostic uses the configured server URL in the kube config to perform a DNS
lookup for the IP(s) of the master. Care is taken to handle both IPv4 and IPv6
addresses being directly in the server URL already.

Once we've got a list of server IPs to check we scan the nodes via the
Kubernetes API, and check all of their reported IP addresses to look for a
match. No DNS lookup is done on the nodes, the IP addresses are already known.

If no node reports the IP of the master server, we warn the user that something
may be amiss and why.

Skip diagnostic if not using an OVS network plugin.

If we have access to a config file, check the network plugin name for one of
the two ovs plugins. If it doesn't match, skip the check.

If no master config file is available (which is common, this is a cluster admin
diagnostic and can be run remotely), run the check anyhow as OVS is the default
today.