Move kube controller base OVS flow setup into setup.sh
multitenant was doing its base OVS flow setup in setup.sh, but kube
was doing part of it in Setup() and part when AddOFRules() got called
for its own IP address. Sync them up.

This also means we no longer remove those rules if we get DelOFRules()
for our own IP address, but in that case we really need a lot more
cleanup than that anyway (eg, the rules added in Setup() never got
removed), so the existing code wasn't really that useful.