Add a LatestBuilds diagnostic.
This diagnostic lists all build configs in the current client context and
project, then checks if the latest build was successful or not. If the build
status reported error or failure we notify the user immediately.

The diagnostic then scans the build logs for any build with failed status, and
applies a series of regexes to each line looking for common problems we could
report on. Today the only problem being detected is the registry being denied
writing to disk due to a common selinux issue.