Rename osdn.go to registry.go, OsdnRegistryInterface to Registry
The file formerly known as "osdn.go" is specifically an implementation
of the openshift-sdn SubnetRegistry, so rename it to "registry.go".

Also, rename "OsdnRegistryInterface" to just "Registry", since the (a)
the package name is "osdn" so that doesn't need to appear in the type
name too, and (b) it's a struct type, not an interface type.