Make deployment trigger logging quieter
Decrease the severity of many common deployment trigger controller
logs which are almost never useful except when debugging in a scenario
where the user has access to the API server and can just increase the
log level and retry for additional details.

Many of these logs are being emitted at V(2) which is extremely spammy
in the master logs for default installations.