Enable binary build endpoint and CLI via start-build
Binary builds are driven by a client launching a build and then
providing a binary input stream for the contents of the build. The build
should accept either a file or archive (tar, tar.gz, or zip) and use the
contents provided in place of normal source download.

The primary REST endpoint is `/buildconfigs/<name>/instantiatebinary`,
which creates a new build from a build config, accepts query parameters
to populate a BinaryBuildRequestOptions object, then on POST creates
the build, waits for it to be running, and attaches to STDIN and streams
the provided input stream to the user.

The start-build command is enhanced to support --from-dir, --from-file,
and --from-repo, which load local directories or files, set source info
(if available) onto the request, and then trigger the build by invoking
the REST endpoint.

The describers in the CLI have been enhanced to show info about binary
builds.