test-cmd.sh: Simplify tests assignment
Fix the code to get the list of tests, and move the assignment of tests
closer to where it is used.

The previous code is problematic for the following reasons:

• The intention of using find is to match recursively.  This intention is
  properly realized.

• The intention of piping the output of the find command into the sort
  command is presumably to produce a sorted list of filenames.  However,
  using the find command with the -print0 flag and the sort command without
  the -z/--zero-terminated flag will result in sort's misinterpreting its
  input, generally interpreting it as a single line (if there are no
  newlines in the filenames printed by find), because sort does not know
  that records are separated by NUL bytes rather than LF bytes.  The sort
  command thus sees its input as a single record and does not actually sort
  its input as intended.

• The intention of using the -print0 flag with find and the -0 flag with
  xargs is presumably to handle filenames that contain whitespace.
  However, the xargs printf command at the end of the pipeline outputs
  records separated by LF bytes, and we use this output in a shell
  substitution, so even if we fixed the problem with sort, we would still
  split filenames on whitespace when we assigned them to the tests
  variable.

It would be possible to rewrite the pipeline as follows to fix the first
problem, so we would sort properly but then split incorrectly on assigning
to tests:

    find "${1}" -name '*.sh' -print0 | sort -zu | xargs -0 -n1 printf "%s\n"

Alternatively, we could forego whitespace-tolerance and simplify the entire
command as follows:

    find "${1}" -name '*.sh' | sort -u

As a further alternative, we could preserve the use of find and sort but
use some more advanced Bash to avoid splitting on whitespace:

    declare -a tests
    while IFS= read -r -d '' filename
    do tests+=("$filename")
    done < <(find "${1}" -name '*.sh' -print0 | sort -zu)

Ultimately, if we want to support filenames that contain whitespace, we
need to do some fairly advanced Bash code.  In this commit, I opt for an
approach that realizes the intentions of recursiveness, sorting the list,
and tolerating whitespace, but that uses only Bash builtins, namely
a direct assignment to the tests array variable using ** for recursive
globbing and relying on the fact that Bash sorts glob matches.