Scan for selinux write error in registry diagnostic.
In the cluster admin diagnostic which scans the registry pod logs, add an
explicit check for an error which indicates the registry cannot write to disk,
which is a strong indicator for an selinux problem. The fix from the origin
troubleshooting page is then suggested.

We also watch for a successful write log entry, and if found *after* the error,
we know the problem has since been fixed and the diagnostic error can be safely
ignored.