Consolidate plugins/osdn and pkg/ovssubnet
No particular reason these are separate, and they share a lot of
code, so consolidate them.

This means we also have to switch to factory functions where the
individual plugins register themselves with the common.go factory
function NewController(), to avoid an import cycle between
pkg/cmd/admin/network, the multiteant plugin, and the core.  But
don't worry, we'll remove that indirection soon.