Prevent dev cluster deploy from using stale config
When a dev cluster is deployed with vagrant, the nodes wait for their
configuration to be created.  If stale configuration is already present,
they may use that configuration before the master has a chance to
generate new configuration.  This change marks existing configuration as
stale at the beginning of provisioning to avoid that possiblity.