Skip sdn node during dev cluster deploy
An sdn node (an openshift node running on the master to ensure master to
pod network connectivity) is a precondition for the kube intra pod test,
but that test doesn't work with origin at present and the method to
disable scheduling of the sdn node has proven unreliable.  This change
makes deployment of an sdn node optional via the OPENSHIFT_SDN_NODE
flag.