Force network tests to wait until cluster is ready
Previously the extended networking tests would be run as soon as dind
cluster creation script returned.  There could be a delay, though, until
nodes in the cluster had registered themselves with the master and were
ready for testing, and if the tests ran before this happened they could
fail due to insufficient nodes being available.  This change ensures
that the tests wait for node availability.