Fix DeploymentConfig Scale Failure Missing Error
When a request to update a DeploymentConfig scale subresource
failed because the corresponding latest deployment was not
"complete", a nil error value would be returned, leading to
a panic (due to Origin trying to serialize a `nil` object to JSON)
and corresponding "Internal Server Error" to be returned to the client.
This commit corrects this so that an actual error is returned by
the Update method.