Create sourcable bash env as part of dind cluster deploy
Deploying a dind cluster on a vm provisioned by vagrant will be
assured of having KUBECONFIG and PATH set correctly in the vm.  For
manual dind deployment (e.g. to an arbitrary docker instance), this
change adds an rc file that can be sourced to configure the
environment for accessing the cluster with the openshift cli.