Fix router panic
If only one route was registered for a host, got deleted, and then
added back again, the cache that ensures unique hosts would panic
due to an index out of range. This commit fixes the panic by ensuring
that when the last route for a host is removed then the entry for that
host is also removed from the cache.