Copy config to dev cluster master's local storage
By default openshift's etcd instance stores data as a peer to the
cluster configuration (e.g. /vagrant/openshift.local.etcd).  When this
path is an nfs mount, various lock-related errors are possible.  This
change copies the configuration to local storage when /vagrant is
mounted over nfs to ensure that etcd's backing store will not be
subject to the possibility of nfs-related lock failures.