Add the ability to disable SSL verification on git over HTTPS
This is done through the new boolean field "disableSSLCheck" inside GitBuildSource
object.

The use case is to allow the user to specify a git repository with a self-signed
SSL certificate. Currently sti build fails at git (ls-remote|clone) stage.

So, one can write a template like this
               ...
                "source": {
                    "contextDir": "${CONTEXT_DIR}",
                    "git": {
                        "disableSSLCheck": true,
                        "ref": "${SOURCE_REPOSITORY_REF}",
                        "uri": "${SOURCE_REPOSITORY_URL}"
                    },
                    "type": "Git"
                ...

In the future, with #3946 implemented, one might use a boolean parameter in the
template and have it rendered into a checkbox in the UI.