Fail early on a node if we can't fetch the ClusterNetwork record
The code was assuming that if the
osClient.ClusterNetwork.Get("default") call in NewRegistry() failed,
it could be ignored, since the corresponding call in StartNode() would
also fail and cause startup to bail out. But apparently it's possible
either that (a) it fails the first time but succeeds the second time,
or (b) it fails but then some other goroutine tries to use the value
of registry.ServiceNetwork before StartNode() has had a chance to
abort.

Unfortunately, the call will *always* fail on the master, because it
gets called before the master is fully up and running. So we have to
split things up so that we just skip the call on the master, but call
it and return an error if it fails on nodes. (Eventually we should do
a better job of splitting up the master and node code in the
plugin...)