Simplify Makefile for limited parallelization
Keep "check" as the "local fast test" and "test" as the "run everything
but extended, assuming Docker".  Remove check-test and some of the other
tests since we aren't using them. Limit TEST_ASSETS to only one spot,
and remove most of the other conditional logic. Remove SKIP_BUILD which
will be replaced with logic in hack/build-go.sh to prefer release tars.
Support parallelization of test-cmd and test-unit for now.