Webhooks: use constant-time string secret comparison
For performance reasons, string comparisons in most languages
short-circuit at the first non-matching character. Unfortunately, when
used for credential verification, this allows an attacker to figure out
the actual credentials by trying values and seeing how long it takes for
the authorization to fail. To avoid this, we use the Go-provided
crypto/hmac.Equal() function, which performs constant-time string
comparison.