Redo OVS rules to avoid ARP caching problems
Acting as a learning switch only works if you're willing to flood
packets that you don't know how to route. OpenShift SDN was doing this
with ARP packets, but not with IP packets, because of isolation. This
*mostly* worked, except that given certain patterns of traffic, we
could end up in a state where container A already knew container B's
MAC address (and so didn't need to send an ARP request), but OVS had
timed out its own MAC-to-port mapping (and couldn't rediscover it
again without seeing an ARP response).

Fix this by not using learn() any more, and just having explicit rules
to direct all of the traffic (like the original single-tenant rules
did...)

Also reorganize things a bit so that we don't need different rules for
single-tenant and multi-tenant (although some of the rules that get
added will never get used in the single-tenant case).